pyproject.toml
[tool.poetry]
name = "localpg"
version = "0.1.0"
description = ""
authors = ["José Castagnino <jose@stamm.bio>"]
packages = [{include = "localpg"}]

[tool.poetry.dependencies]
python = "3.8.10"
pandas = "^1.3.5"
alembic = "^1.7.5"
muttlib = "1.4.12"
python-decouple = "^3.5"
pangres = "^3.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
pdbpp = "^0.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
